refactor dtos, mappers, model structure #JC19-359
