+++ apps/vanilla/src/scripts/animeTable.ts
@@ -2,7 +2,9 @@ import { Anime } from '@js-camp/core/models/anime';
 import { Pagination } from '@js-camp/core/models/pagination';
 import { assertNonNull } from '@js-camp/core/utils/assertNonNull';
 
-import { dateConverter } from '../scripts/functions';
+import { dateConverter, navigate } from '../scripts/functions';
+
+import { Url } from './constants';
 
 import { renderPaginateButton } from './pagination';
 import { LIMIT } from './variables';
@@ -15,7 +17,7 @@ export function renderAnimeTable(dataset: Pagination<Anime>): void {
   const dataTable = document.querySelector<HTMLDivElement>('.anime-table');
   assertNonNull(dataTable);
   const tableBody = dataset.results.reduce((previousValue: string, currentValue: Anime): string => `${previousValue}
-    <tr>
+    <tr data-id = "${currentValue.id}">
          <td>
            <img class="anime-image" src="${currentValue.image}" alt="Anime image">
          </td>
@@ -46,4 +48,14 @@ export function renderAnimeTable(dataset: Pagination<Anime>): void {
   `;
   const totalPages = dataset.count / LIMIT;
   renderPaginateButton(totalPages);
+  const animeTableRows = document.querySelectorAll<HTMLTableRowElement>('tbody tr');
+  assertNonNull(animeTableRows);
+  animeTableRows.forEach(anime => {
+    anime.addEventListener('click', () => {
+      const animeId = anime.getAttribute('data-id');
+      assertNonNull(animeId);
+      localStorage.setItem('CLICKED_ANIME_ID', animeId);
+      navigate(Url.Detail);
+    });
+  });
 }
