+++ libs/core/dtos/anime/genre.dto.ts
