+++ libs/core/dtos/auth/register.dto.ts
