+++ libs/core/mappers/auth/register.mapper.ts
@@ -1,5 +1,5 @@
-import { RegisterDto } from '../dtos/register.dto';
-import { Register } from '../models/register';
+import { RegisterDto } from '../../dtos/auth/register.dto';
+import { Register } from '../../models/register';
 
 export namespace RegisterMapper {
